requirements.txt
openai-whisper
google-generativeai
pydub
PyMuPDF
python-pptx
python-dotenv
numpy
soundfile
librosa
